ref: Reuse ByteArray.toHex().
