+++ src/main/kotlin/org/jitsi/rtp/extensions/ByteArrayBuffer.kt
@@ -16,30 +16,8 @@
 
 package org.jitsi.rtp.extensions
 
+import org.jitsi.rtp.extensions.bytearray.toHex
 import org.jitsi.utils.ByteArrayBuffer
 
 @JvmOverloads
-fun ByteArrayBuffer.toHex(maxBytes: Int = Int.MAX_VALUE): String {
-    val HEX_CHARS = "0123456789ABCDEF"
-
-    val numBytes = minOf(maxBytes, length, buffer.size - offset)
-
-    val sb = StringBuilder()
-    for (i in offset until offset + numBytes) {
-        val position = i - offset
-        if (position != 0) {
-            if (position % 16 == 0) {
-                sb.append("\n")
-            } else if (position % 4 == 0) {
-                sb.append(" ")
-            }
-        }
-        val byte: Int = buffer[i].toInt()
-        val firstIndex = (byte and 0xF0) shr 4
-        val secondIndex = byte and 0x0F
-        sb.append(HEX_CHARS[firstIndex])
-        sb.append(HEX_CHARS[secondIndex])
-    }
-
-    return sb.toString()
-}
+fun ByteArrayBuffer.toHex(maxBytes: Int = Int.MAX_VALUE) = buffer.toHex(offset, maxBytes.coerceAtMost(length))
